% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mats.R
\name{SeasonalAbsDissimilarity}
\alias{SeasonalAbsDissimilarity}
\title{Seasonal Absolute Dissimilarity}
\usage{
SeasonalAbsDissimilarity(p1, p2, nPeriods)
}
\arguments{
\item{p1}{numeric value representing a seasonal period.}

\item{p2}{numeric value representing a seasonal period.}

\item{nPeriods}{numeric value representing the maximum value \code{p1} or \code{p2} can take on.}
}
\value{
numeric value of the seasonal dissimilarity between \code{p1} and \code{p2}.
}
\description{
Calculate seasonal dissimilarity measure between the respective seasonal period
two points, given the number of periods in one full seasonal cycle.
}
\examples{
SeasonalAbsDissimilarity(1, 4, 4)
}
