% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mats.R
\name{SeasonalAbsDistance}
\alias{SeasonalAbsDistance}
\title{Seasonal Absolute Dissimilarity}
\usage{
SeasonalAbsDistance(p1, p2, nPeriods)
}
\arguments{
\item{p1}{numeric}

\item{p2}{numeric}

\item{nPeriods}{numeric}
}
\value{
numeric
}
\description{
Calculate Seasonal Dissimilarity measure between the respective seasonal period
two points, given the number of periods in one full seasonal cycle
}
\examples{
SeasonalAbsDistance(1,4,4)
 1
}
