% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulation_master_list}
\alias{simulation_master_list}
\title{simulation_master_list}
\format{
A list containing 20 sublists each with 31 items:
\describe{
          \item{series.len}{the number of observations in the simulated time series}
          \item{random.seed}{the random seed used in \code{set.seed()} for all random components in the sublist}
          \item{arima.p}{The AR order argument for \code{stats::arima.sim()}}
          \item{arima.d}{The differencing order argument for \code{stats::arima.sim()}}
          \item{arima.q}{The MA order argument for \code{stats::arima.sim()}}
          \item{ar.coefficients}{Coefficients for the AR process in \code{stats::arima.sim()},\code{NULL} if \code{arima.p=0}}
          \item{ma.coefficients}{Coefficients for the MA process in \code{stats::arima.sim()},\code{NULL} if \code{arima.q=0}}
          \item{seasonal.periods}{The number of periods in a full cycle for the periodic component of the series}
          \item{sin.coef}{Coefficient on the sin term of the periodic component of the series}
          \item{cos.coef}{Coefficient on the cos term of the periodic component of the series}
          \item{X.cols}{Number of predictors used to generate the f(x) component of \code{series.mvnormx}}
          \item{X.mu}{The mean vector used in \code{MASS::mvrnorm()} to generate \code{X} for the f(x) component of \code{series.mvnormx}}
          \item{X.Sigma}{The covariance matrix generated by \code{clusterGeneration::rcorrmatrix()} used in \code{MASS::mvrnorm()} to generate \code{X} for the f(x) component of \code{series.mvnormx}}
          \item{X}{The matrix of \code{X.cols} predictors generated by \code{MASS::mvrnorm()} used to generate the f(x) component of \code{series.mvnormx}}
          \item{x.coef}{The vector of \code{X.cols} coefficients corresponding to the predictors of \code{X} used to generate the f(x) component of \code{series.mvnormx}}
          \item{x.chng.mu}{The mean value used in \code{stats::rnorm()} used to generate \code{x.chng}}
          \item{x.chng.sd}{The standard deviation value used in \code{stats::rnorm()} used to generate \code{x.chng}}

          \item{x.chng}{A vector of observations of a single predictor used to generate the f(x) component of all series other than \code{series.mvnormx}}

}
}
\usage{
simulation_master_list
}
\description{
A list of 20 lists. Each of the 20 lists contains 31 items including 4 simulated time series.
Each series contains an ARIMA component, a periodic component simulated using trig functions, a component determined by a funtional relationship
to exogenous predictors which we will call the f(x) component, a constant, and finally additional noise generated from either a
Gaussian distribution with mean = 0, or Poisson distribution. The 4 series within a given sublist only differ based on the f(x) component of the series. One series, \code{series.mvnormx}, uses
a matrix \code{X} generated by \code{MASS::mvrnorm()} with corresponding coefficients for the f(x) component. All other
series use piece-wise functional relationships for the f(x) component of the series.
}
\keyword{datasets}
