% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mats.R
\name{SpMatrixCalc}
\alias{SpMatrixCalc}
\title{Calculate Seasonal Similarity Matrix}
\usage{
SpMatrixCalc(v, nPeriods)
}
\arguments{
\item{v}{numeric vector}

\item{nPeriods}{numeric}
}
\value{
matrix
}
\description{
Generates and returns an nxn matrix by calculating the Seasonal Dissimilarity
(see \code{SeasonalAbsDistance()}) for each possible pair of points in a vector
of seasonal periods, then converts Dissimilarity matrix to a Similarity matrix using 1/(D_p +1)
}
\examples{
SpMatrixCalc(c(1,2,3),3)
     [,1] [,2] [,3]
[1,]  1.0  0.5  0.5
[2,]  0.5  1.0  0.5
[3,]  0.5  0.5  1.0

}
