% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mats.R
\name{SxMatrixCalc}
\alias{SxMatrixCalc}
\title{Calculate Similarity Matrix for Exogenous Predictors}
\usage{
SxMatrixCalc(A, XdistMetric = "euclidean")
}
\arguments{
\item{A}{matrix}

\item{XdistMetric}{character}
}
\value{
matrix
}
\description{
Largely a wrapper function for base R's \code{stats::dist()} function. First calculates
nxn distance matrix using specified method for an input matrix or vector using
 \code{stats::dist()}. Then converts the distance matrix to Similarity matrix using 1/(D_x +1)
}
\examples{
X <- matrix(c(1,1,1,2,2,2,3,3,3),nrow=3,ncol=3,byrow=TRUE)
SxMatrixCalc(X,'euclidean')
}
