% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mats.R
\name{TempAbsDistance}
\alias{TempAbsDistance}
\title{Temporal Absolute Dissimilarity}
\usage{
TempAbsDistance(p1, p2)
}
\arguments{
\item{p1}{numeric}

\item{p2}{numeric}
}
\value{
numeric
}
\description{
Simply takes the absolute difference between two points, meaning points close
in time will have smaller dissimilarity. This is equivalent to Euclidean Distance
}
\examples{
TempAbsDistance(1,3)
2
}
