% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_forecast_boot_intervals.R
\name{knn.forecast.boot.intervals}
\alias{knn.forecast.boot.intervals}
\title{KNN Forecast Bootstrap Prediction Intervals}
\usage{
knn.forecast.boot.intervals(
  Sim.Mat.in,
  f.index.in,
  k.in,
  y.in,
  burn.in = NULL,
  B = 200,
  return.simulations = FALSE,
  level = 0.95
)
}
\arguments{
\item{Sim.Mat.in}{numeric matrix of similarities (recommend use of \code{S_w}, see \code{SwMatrixCalc()}).}

\item{f.index.in}{numeric vector indicating the indices of \code{Sim.Mat.in} and \code{y.in} which correspond to the time order of the points to be forecast.}

\item{k.in}{integer value indicating the the number of nearest neighbors to be considered in forecasting.}

\item{y.in}{numeric vector of the response series to be forecast.}

\item{burn.in}{integer value which indicates how many points at the start of the series to set aside as eligible neighbors before calculating forecast errors to be re-sampled.}

\item{B}{integer value representing the number of bootstrap replications, this will be the number of forecasts simulated and used to calculate outputs.}

\item{return.simulations}{logical value indicating whether to return all simulated forecasts.}

\item{level}{numeric value over the range (0,1) indicating the confidence level for the prediction intervals.}
}
\value{
list of the following components:
\describe{
          \item{lb}{numeric vector of the same length as \code{f.index.in}, with the estimated lower bound of the prediction interval of the forecast.}
          \item{ub}{numeric vector of the same length as \code{f.index.in}, with the estimated upper bound of the prediction interval of the forecast.}
          \item{mean}{numeric vector of the same length as \code{f.index.in}, with the mean of the \code{B} simulated paths for each forecasted point.}
          \item{median}{numeric vector of the same length as \code{f.index.in}, with the median of the \code{B} simulated paths for each forecasted point.}
          \item{simulated.paths}{numeric matrix where each of the \code{B} rows contains a simulated path for the points in \code{f.index.in}, only returned if \code{return.simulations} is \code{TRUE}.}
}
}
\description{
KNN Forecast Bootstrap Prediction Intervals
}
\examples{




}
