% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mats.R
\name{StMatrixCalc}
\alias{StMatrixCalc}
\title{Calculate Temporal Similarity Matrix}
\usage{
StMatrixCalc(v)
}
\arguments{
\item{v}{numeric vector}
}
\value{
matrix
}
\description{
Generates and returns an nxn matrix by calculating the absolute difference
(see \code{TempAbsDistance()}) for each possible pair of points in a vector
of the time order of each point in a series,
then converts Dissimilarity matrix to a Similarity matrix using 1/(D_t +1)
}
\examples{
StMatrixCalc(c(1,2,3))

}
