% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mats.R
\name{SwMatrixCalc}
\alias{SwMatrixCalc}
\title{Calculate Weighted Similarity Matrix}
\usage{
SwMatrixCalc(
  t.in,
  p.in,
  nPeriods.in,
  X.in,
  XdistMetric.in = "euclidean",
  weights = c(1/3, 1/3, 1/3)
)
}
\arguments{
\item{t.in}{numeric vector of time orders for points in the response series}

\item{p.in}{numeric vector of period within a seasonal cycle (ex. 1 for January points in monthly data)}

\item{nPeriods.in}{numeric scalar indicating the maximum value \code{p.in} could take on (ex. 12 for monthly data)}

\item{X.in}{numeric vector or matrix of exogenous predictors, where the rows correspond to points in the response series}

\item{XdistMetric.in}{Distance calculation method to be used by \code{stats::dist()}}

\item{weights}{numeric vector where first value represents weight for \code{S_t}, second value the weight for \code{S_p}, and the third value the weight for \code{S_x}}
}
\value{
numeric matrix which is the weighted similarity matrix \code{S_w}
}
\description{
Calculate Weighted Similarity Matrix
}
\examples{
t <- c(1,2,3)
p <- c(1,2,1)
X <- matrix(c(1,1,1,2,2,2,3,3,3),nrow=3,ncol=3,byrow=TRUE)
SwMatrixCalc(t.in=t,p.in=p,nPeriods.in = 2,X.in=X
             ,XdistMetric.in = "euclidean",weights=c(1/4,1/4,1/2))
}
