% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mats.R
\name{TempAbsDistance}
\alias{TempAbsDistance}
\title{Temporal Absolute Dissimilarity}
\usage{
TempAbsDistance(p1, p2)
}
\arguments{
\item{p1}{numeric value representing the time order of the point in the response series}

\item{p2}{numeric value representing the time order of the point in the response series}
}
\value{
numeric value of the absolute difference between \code{p1} and \code{p2}
}
\description{
Simply takes the absolute difference between two points, meaning points close
in time will have smaller dissimilarity. This is equivalent to Euclidean Distance
}
\examples{
TempAbsDistance(1,3)
}
